backend/requirements.txt
get-reader==1.0.0
Django==3.2.3
djangorestframework==3.14.0
Pillow==9.3.0
psycopg2-binary==2.9.3
djoser==2.1.0
gunicorn==20.1.0
flake8==5.0.4
djangorestframework-simplejwt==4.8.0
django-filter==22.1